Issue #3118987 by chmez: Cancel changes for replacing select field by checkbox field
