+++ modules/social_features/social_content_block/config/install/core.entity_view_display.block_content.custom_content_list.default.yml
@@ -5,7 +5,7 @@ dependencies:
     - block_content.type.custom_content_list
     - field.field.block_content.custom_content_list.field_item_amount
     - field.field.block_content.custom_content_list.field_link
-    - field.field.block_content.custom_content_list.field_plugin_filters
+    - field.field.block_content.custom_content_list.field_plugin_field
     - field.field.block_content.custom_content_list.field_plugin_id
     - field.field.block_content.custom_content_list.field_sorting
     - field.field.block_content.custom_content_list.field_subtitle
@@ -16,13 +16,13 @@ mode: default
 content:
   entities:
     weight: 0
-    region: content
     settings: {  }
     third_party_settings: {  }
+    region: content
 hidden:
   field_item_amount: true
   field_link: true
-  field_plugin_filters: true
+  field_plugin_field: true
   field_plugin_id: true
   field_sorting: true
   field_subtitle: true
