+++ modules/social_features/social_content_block/config/install/field.storage.block_content.field_plugin_field.yml
@@ -3,11 +3,14 @@ status: true
 dependencies:
   module:
     - block_content
-id: block_content.field_plugin_filters
-field_name: field_plugin_filters
+id: block_content.field_plugin_field
+field_name: field_plugin_field
 entity_type: block_content
-type: boolean
-settings: {  }
+type: string
+settings:
+  max_length: 255
+  is_ascii: false
+  case_sensitive: false
 module: core
 locked: false
 cardinality: 1
