+++ modules/social_features/social_content_block/social_content_block.install
@@ -381,17 +381,3 @@ function social_content_block_update_8005() {
   return $update_helper->logger()->output();
 }
 
-/**
- * Configuration update.
- */
-function social_content_block_update_8006() {
-  /** @var \Drupal\update_helper\Updater $updateHelper */
-  $updateHelper = \Drupal::service('update_helper.updater');
-
-  // Execute configuration update definitions with logging of success.
-  $updateHelper->executeUpdate('social_content_block', 'social_content_block_update_8006');
-
-  // Output logged messages to related channel of update execution.
-  return $updateHelper->logger()->output();
-}
-
